Pulled PlayerStatuses out of GameState
